Regenerate yarn.lock
The package.json file lists direct dependencies of our
project. Dependabot automatically updates this list and regenerates
the yarn.lock based on these updates. With this process in place, all
direct dependencies are always up to date and this is great.

The problem is with indirect dependencies. Imagine that our
package.json declares a dependency towards SuperTool at version
"^1.0.0". Imagine now that SuperTool's own package.json declares a
dependency towards SuperLib at version "^1.0.0". At this point, if
both projects latest release is 1.0.0, our yarn.lock contains a
reference to both SuperTool@1.0.0 and SuperLib@1.0.0. If SuperLib
later gets a new release 1.1.0, we will never get it because (1)
dependabot won't bother upgrading an indirect dependency and (2)
yarn.lock mentions explicitly that we want version 1.0.0.

From time to time, it is good to delete the yarn.lock file and ask
yarn to generate it from scratch. In our example above, this will
write "version 1.1.0" for SuperLib into yarn.lock because this is the
latest release of the package that is compatible with SuperTool.